Use `StringColumn` for decimal type by default (#509)

+++ tests/Stubs/ActiveRecord/Customer.php
@@ -21,7 +21,7 @@ class Customer extends ArrayableActiveRecord
     public const STATUS_INACTIVE = 2;
 
     public int|string $status2;
-    public int|string|null $sumTotal;
+    public float|string|null $sumTotal;
 
     protected int $id;
     protected string $email;
